Rename Display to Project
